Path gets  a visitIdentifiers method
